+++ persistence/src/test/resources/SQL/test_db.sql
@@ -139,7 +139,7 @@ create table if not exists orders
     user_id int not  null,
     certificate_id int not null,
     purchase_date timestamp not null,
-    amount float not null,
+    amount double not null,
     quantity    int not null,
     foreign key(user_id) references user(id) on update cascade  on delete cascade,
     foreign key(user_id) references account(id)  on update cascade on delete cascade,
